UV LOCATION INDEXES RELEASED
Over the past few days, myself and multiple other people working on liveries have frequently gotten stuck on finding which part of the UV map or 3D model corresponds to which part of the texture, so I spent some time designing these UV map grids. Essentially, find the part on the aircraft below that you want to texture, identify which square covers that part, and find it on the UV map overlay above to identify which part of the UV map corresponds to which part. Hope it helps!